PR Comment Fixed
- made new folder for ViewModifier and moved TrailingFadeGradient to withTrailingFadeGradient
- renamed TextUtil file to String+Extention
